+++ keyboards/handwired/polykybd/poly_util.h
@@ -9,7 +9,7 @@ void select_all_displays(void);
 
 void clear_all_displays(void);
 
-void display_message(uint8_t row, uint8_t col, const uint16_t* message, const GFXfont* font);
+void display_message(uint8_t row, uint8_t col, const uint32_t* message, const GFXfont* font);
 
 void display_bootloader_message(void);
